new schema changes and some refactoring
