+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@adobe/data",
-  "version": "0.5.22",
+  "version": "0.5.23",
   "description": "Adobe data oriented programming library",
   "type": "module",
   "sideEffects": false,
