--- src/ecs/store/store-schema/create-store-schema.ts
@@ -1,50 +1,0 @@
-/*MIT License
-
-© Copyright 2025 Adobe. All rights reserved.
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.*/
-
-import { StringKeyof } from "../../../types/types.js";
-import { ComponentSchemas } from "../../component-schemas.js";
-import { ResourceSchemas } from "../../resource-schemas.js";
-import { ArchetypeComponents } from "../archetype-components.js";
-import { createStore } from "../create-store.js";
-import { StoreSchema } from "./store-schema.js";
-
-export function createStoreSchema<
-    const CS extends ComponentSchemas,
-    const RS extends ResourceSchemas,
-    const A extends ArchetypeComponents<StringKeyof<CS>>,
->(
-    components: CS,
-    resources: RS,
-    archetypes: A,
-) {
-    return { components, resources, archetypes } as const satisfies StoreSchema<CS, RS, A>;
-};
-
-export function createStoreFromSchema<
-    const CS extends ComponentSchemas,
-    const RS extends ResourceSchemas,
-    const A extends ArchetypeComponents<StringKeyof<CS>>,
->(
-    schema: StoreSchema<CS, RS, A>,
-) {
-    return createStore(schema.components, schema.resources, schema.archetypes);
-};
