--- src/ecs/store/store-schema/store-schema.ts
@@ -1,63 +1,0 @@
-/*MIT License
-
-© Copyright 2025 Adobe. All rights reserved.
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.*/
-
-import { FromSchemas } from "../../../schema/schema.js";
-import { ComponentSchemas } from "../../component-schemas.js";
-import { ResourceSchemas } from "../../resource-schemas.js";
-import { Store } from "../store.js";
-import { ArchetypeComponents } from "../archetype-components.js";
-import { StringKeyof } from "../../../types/types.js";
-import { Assert } from "../../../types/assert.js";
-import { Equal } from "../../../types/equal.js";
-import { Archetype } from "../../archetype/archetype.js";
-import { CoreComponents } from "../../core-components.js";
-
-export type StoreSchema<
-    CS extends ComponentSchemas,
-    RS extends ResourceSchemas,
-    A extends ArchetypeComponents<StringKeyof<CS>>,
-> = {
-    readonly components: CS;
-    readonly resources: RS;
-    readonly archetypes: A;
-};
-
-export type StoreFromSchema<T> = T extends StoreSchema<infer CS, infer RS, infer A> ? Store<FromSchemas<CS>, FromSchemas<RS>, A> : never;
-
-type CheckStoreFromSchema = StoreFromSchema<StoreSchema<{
-    velocity: { type: "number" },
-    particle: { type: "boolean" },
-}, {
-    mousePosition: { type: "number", default: 0 },
-    fooPosition: { type: "number", default: 0 },
-}, {
-    Particle: ["particle"],
-    DynamicParticle: ["particle", "velocity"],
-}>>;
-declare const testStore: CheckStoreFromSchema;
-type CheckDynamicParticle = Assert<Equal<typeof testStore.archetypes.DynamicParticle, Archetype<CoreComponents & {
-    particle: boolean;
-    velocity: number;
-}>>>;
-type CheckParticle = Assert<Equal<typeof testStore.archetypes.Particle, Archetype<CoreComponents & {
-    particle: boolean;
-}>>>;
